+++ mlrun/api/api/endpoints/feature_sets.py
@@ -59,7 +59,7 @@ def patch_feature_set(
     feature_set_update: dict,
     reference: str,
     patch_mode: schemas.PatchMode = Header(
-        schemas.PatchMode.replace, alias="x-mlrun-patch-mode"
+        schemas.PatchMode.replace, alias=schemas.HeaderNames.patch_mode
     ),
     db_session: Session = Depends(deps.get_db_session),
 ):
